% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_traj1_set.R
\name{create_traj1_set}
\alias{create_traj1_set}
\title{Create standardized trajectory object}
\usage{
create_traj1_set(
  data,
  nobs_persub = getOption("trajclass.nobs_persub", default = 10),
  .msg = TRUE
)
}
\arguments{
\item{data}{A data frame or tibble with longitudinal data}

\item{nobs_persub}{Minimum number of non-missing observations per subject
(default = 5). Subjects with fewer valid rows are excluded.}

\item{key_vars}{\strong{Required.} Named character vector of length 3 defining the column names
Must have names \code{"id_col"}, \code{"time_col"}, \code{"y_col"}
Example: \code{c(id_col = "ID", time_col = "time", y_col = "egfr")}}
}
\value{
A tibble with standardized columns \code{"id"}, \code{"time"}, \code{"y"} and class:
\itemize{
\item \verb{"traj1" "traj"} — exactly one subject
\item \code{"traj"} — multiple subjects or empty
}

Attribute \code{"data_info"} is a tibble containing two rows
}
\description{
Prepares a clean tibble with standardized columns \code{id}, \code{time}, \code{y}
from a longitudinal dataset (stored in \code{traj} object. Applies the specified filter (default: ≥5 non-NA \code{y} values per subject).

The returned object always has base class \code{"traj"}, and adds subclass \code{"traj1"}
if exactly one unique subject remains after filtering.

Metadata is attached as attribute \code{"data_info"} (a tibble).
}
\examples{
\dontrun{
orig_keys <- getOption("trajclass.key_vars", default = NULL)
traj_df <- create_traj_data(example_data, key_vars = orig_keys)
traj1_set <- create_traj1_set(traj_df)
class(traj1_set)
length(traj1_set)                  # number of selected subjects
 head(names(traj1_set))             # "1_10" "1_11" "1_15" ... 
traj1 <-  traj1_set[["1_15"]]                    
attrx <- attributes(traj1)
names(attrx)  #   "names"  "row.names"  "class"      "traj1_info"
attrx$names   #  "id"   "time" "y" 
head(attrx$row.names)   # [1] 1 2 3 4 5 6
}
}
