% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_traj_data.R
\name{create_traj_data}
\alias{create_traj_data}
\title{Create a base trajectory data object (pre-processing step)}
\usage{
create_traj_data(
  data,
  key_vars = getOption("trajclass.key_vars", default = NULL),
  .msg = TRUE
)
}
\arguments{
\item{data}{A data frame or tibble in long format}

\item{key_vars}{\strong{Required.} Named character vector of length 3 defining the column names.
Must have names \code{"id_col"}, \code{"time_col"}, \code{"y_col"}.
Example: \code{c(id_col = "ID", time_col = "time", y_col = "egfr")}}
}
\value{
A tibble with names of key columns columns retained  (object of class \verb{"traj_data")}
(inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}).
All rows with missing values in any key column are removed.

Attributes:
\itemize{
\item \code{data_info} : tibble with one row describing input data
}
}
\description{
Takes original long-format data, selects the three key columns specified in \code{key_vars},
removes any row where \strong{any} of the three key columns is missing (\code{NA})

The result is a minimal tibble with class \code{"traj_data"} (inheriting from \code{tbl_df} / tibble).
One attribute is added:
\itemize{
\item \code{data_info} — tibble with one row with basic information on input data
}

This function is intended as the very first pre-processing step before further
operations (filtering by nobs per subject, classification, etc.).
}
\examples{
\dontrun{
# Basic usage with example_data
orig_keys <- getOption("trajclass.key_vars", default = NULL)
traj_df <- create_traj_data(example_data, key_vars = orig_keys)


# Check attributes
attrx <- attributes(traj_df)
names(attrx)                      # "names" "row.names" "class" "info"
names(attrx$info)                 # "orig_key_vars" "datain" "dataout" 
attrx$info$orig_key_vars          # named list
attrx$info$datain                 # a tibble with 1 row
attrx$info$dataout$key_vars       # character vector with names, e.g. [1] "ID"   "time" "egfr"
# Check if No rows with missing key values
any(is.na(traj_df$ID))          # FALSE
any(is.na(traj_df$time))        # FALSE
any(is.na(traj_df$egfr))        # FALSE 

}

}
