% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{inspect_parent_data}
\alias{inspect_parent_data}
\title{Inspect 'data' argument of the calling function and return info as tibble}
\usage{
inspect_parent_data(.msg = FALSE)
}
\arguments{
\item{.msg}{logical. Whether to print a short message about the data
(default: \code{TRUE})}
}
\value{
A one-row \code{tibble} with information about the \code{data} argument
}
\description{
Inspect 'data' argument of the calling function and return info as tibble
}
\details{
Must be called as the \strong{very first expression} inside a function
that has a formal argument named \code{data}.
}
