% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_augment_fit_method.R
\name{my_augment}
\alias{my_augment}
\title{Tidy augment tibble forsingle-model objects}
\usage{
my_augment(x, se_fit = TRUE, ...)
}
\arguments{
\item{x}{a single fitted model extracted from traj_fit object}

\item{...}{Passed to \code{parameters::model_parameters()}}

\item{ci}{Confidence level for intervals (default 0.95)}
}
\value{
A tibble with model coefficients (one row per term)
}
\description{
Tidy augment tibble forsingle-model objects
}
