% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_glimpse_lm_method.R
\name{my_glimpse.lm}
\alias{my_glimpse.lm}
\title{Glimpse (one-row summary) for lm / segmented models derived from traj1_fit}
\usage{
\method{my_glimpse}{lm}(x, ...)
}
\arguments{
\item{x}{A fitted model object inheriting from \code{"lm"} (possibly with \code{"traj1_fit"} ancestry)}

\item{...}{Passed to other extractors (not currently used)}
}
\value{
A one-row tibble with key model summary statistics
}
\description{
Glimpse (one-row summary) for lm / segmented models derived from traj1_fit
}
