% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_glimpse_lm_method.R
\name{my_glimpse}
\alias{my_glimpse}
\title{Tidy glimpse tibble for single-model objects}
\usage{
my_glimpse(x, ...)
}
\arguments{
\item{x}{a single fitted model extracted from traj_fit object}

\item{...}{Passed to \code{parameters::model_parameters()}}
}
\value{
A tibble with one row
}
\description{
Tidy glimpse tibble for single-model objects
}
