% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runscript.R
\name{runScript}
\alias{runScript}
\title{Execute scripts from trajclass package}
\usage{
runScript(
  script = NA,
  package = "trajclass",
  subdir = "scripts",
  echo = TRUE,
  max.deparse = 10000
)
}
\arguments{
\item{script}{character string containing the name of the script to be executed. By default, \code{NA}.}

\item{package}{character string containing the package name. By default, \code{"trajclass"}.}

\item{subdir}{subdirectory containing scripts. By default, \code{"scripts"}.}

\item{echo}{logical. If \code{TRUE}, the script is executed with output printed. Used by \code{source()}. By default, \code{TRUE}.}
}
\value{
The script is executed, and results are printed. If \code{script} is \code{NA}, a list of available scripts is printed.
}
\description{
Executes scripts from trajclass package. If called without arguments, it prints a list of available scripts.
}
\examples{
runScript()
}
