% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runscript_with_log.R
\name{runScript_with_log}
\alias{runScript_with_log}
\title{runScript_with_log}
\usage{
runScript_with_log(
  script_name,
  scriptsDir = NULL,
  log_file = NULL,
  split = TRUE,
  max_deparse = 10000
)
}
\description{
runScript_with_log
}
