% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_traj1_fit.R
\name{select_traj1_fit}
\alias{select_traj1_fit}
\title{Select one model fit from a traj1_fit object}
\usage{
select_traj1_fit(x, idx = 1L)
}
\arguments{
\item{x}{An object of class \code{"traj1_fit"}}

\item{idx}{Integer (1, 2, 3) or \code{NA}:
\itemize{
\item 1 = linear
\item 2 = quadratic
\item 3 = hockey-stick
\item NA = return empty object
}}
}
\value{
A list with class vector starting with \code{"lin"}, \code{"quad"}, or \code{"hstick"},
followed by \code{"traj1_fit"} and \code{"list"}. Retains original attributes.
}
\description{
Select one model fit from a traj1_fit object
}
