% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_traj.R
\name{subset_traj}
\alias{subset_traj}
\alias{subset_traj.traj_0}
\alias{subset_traj.traj_1}
\alias{subset_traj.traj_m}
\title{Subset a traj object}
\usage{
subset_traj(x, ids = NULL)

\method{subset_traj}{traj_0}(x, ids = NULL)

\method{subset_traj}{traj_1}(x, ids = NULL)

\method{subset_traj}{traj_m}(x, ids = NULL)
}
\arguments{
\item{x}{An object of class \code{"traj"} or subclass}

\item{ids}{Character vector of subject IDs to retain
\code{NULL} → select only the first subject (if any)}
}
\value{
Object of the same class family as \code{x} (or \code{"traj_0"} if empty)
}
\description{
Generic function for subsetting objects inheriting from \code{"traj"}.
Dispatches to specific methods based on the subclass (\code{traj_0}, \code{traj_1}, \code{traj_m}).
}
