% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajclass_options.R
\name{trajclass_options}
\alias{trajclass_options}
\title{View current trajclass options (changed or all)}
\usage{
trajclass_options(all = FALSE)
}
\arguments{
\item{all}{If \code{TRUE}, show all trajclass.* options. If \code{FALSE} (default),
show only options that differ from the package defaults.}
}
\value{
Named list of current options
}
\description{
View current trajclass options (changed or all)
}
